## Description

<!-- Add a more detailed description of the changes if needed. -->

Runtime performance of saving a feature by issuing `pytest -vv tests/unit/api/test_feature.py -k test_feature_or_view_column_name_contains_quote -s ` (using this branch: https://github.com/featurebyte/featurebyte/tree/senbong/print-node).
```
----> Reconfiguring loggers in get_client: 0.020622730255126953 seconds
----> Quick prune for payload: 0.001547098159790039 seconds
----> Reconfiguring loggers in get_client: 0.02048182487487793 seconds
----> Prepare task payload (api server): 0.001299142837524414 seconds
----> Quick prune for definition (BatchFeatureCreate.create_feature): 0.0013282299041748047 seconds
------> View construction (prepare_graph_to_store): 0.01244974136352539 seconds
------> Prune graph & node parameters (prepare_graph_to_store): 0.0038080215454101562 seconds
------> Sanitize query graph for definition (prepare_graph_to_store): 0.00115203857421875 seconds
------> Extract derived attributes (service.feature): 0.002209901809692383 seconds
------> Generate definition (namespace_handler.prepare_definition): 0.01884293556213379 seconds
----> Prune for definition (BatchFeatureCreate.create_feature): 0.04477190971374512 seconds
------> Reconfiguring loggers in get_client: 0.020976781845092773 seconds
------> Reconfiguring loggers in get_client: 0.02047896385192871 seconds
------> Reconfiguring loggers in get_client: 0.020095109939575195 seconds
--------> Add GroupBy node to global graph: 0.003368854522705078 seconds
------> Aggregate over operation: 0.0036559104919433594 seconds
----------> Reconfiguring loggers in get_client: 0.020027875900268555 seconds
--------> Check object not been saved (savable_api_object): 0.023073196411132812 seconds
----------> Reconfiguring loggers in get_client: 0.020377159118652344 seconds
--------> Get client (savable_api_object): 0.020481109619140625 seconds
--------> Create payload for feature (api.feature: 0.003451824188232422 seconds
------> Save feature (server mode, api.feature): 0.047338008880615234 seconds
----> Execute SDK code (BatchFeatureCreate.create_feature): 0.1287243366241455 seconds
------> View construction (prepare_graph_to_store): 0.012315034866333008 seconds
------> Prune graph & node parameters (prepare_graph_to_store): 0.003863096237182617 seconds
------> Extract derived attributes (service.feature): 0.0022399425506591797 seconds
------> Generate definition (namespace_handler.prepare_definition): 0.016345977783203125 seconds
------> Generate definition hash (service.feature): 0.001428842544555664 seconds
----> Create feature (controller): 0.08942484855651855 seconds
----> Reconfiguring loggers in get_client: 0.020383119583129883 seconds
----> Reconfiguring loggers in get_client: 0.02598094940185547 seconds
--> Saving feature: 0.511566162109375 seconds
```
After making this change, the runtime of saving a feature is reduce to about 0.46s.

## Related Issue

<!-- If your PR refers to a related issue, link it here. -->

## Type of Change

<!-- Mark with an `x` all the checkboxes that apply (like `[x]`) -->

Label this pull request to place it under the correct category in Release Notes:

|        **Pull Request Label**         | **Category in Release Notes** |
|:-------------------------------------:|:-----------------------------:|
|       `enhancement`, `feature`        |          🚀 Features          |
| `bug`, `refactoring`, `bugfix`, `fix` |    🔧 Fixes & Refactoring     |
|       `build`, `ci`, `testing`        |    📦 Build System & CI/CD    |
|              `breaking`               |      💥 Breaking Changes      |
|            `documentation`            |       📝 Documentation        |
|            `dependencies`             |    ⬆️ Dependencies updates    |



## Checklist

<!-- Mark with an `x` all the checkboxes that apply (like `[x]`) -->

- [ ] I have read the [`CODE_OF_CONDUCT.md`](https://github.com/featurebyte/featurebyte/blob/main/CODE_OF_CONDUCT.md) and [`CONTRIBUTING.md`](https://github.com/featurebyte/featurebyte/blob/main/CONTRIBUTING.md) guides.
- [ ] I have written tests for the changes made.
- [ ] I have written docstrings in [NumpyDoc format](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard)
- [ ] I have labeled my Pull Request correctly